+++ R/encoding.R
@@ -133,7 +133,7 @@
 #' is a translation scheme: we need to communicate with \R somehow,
 #' relying on how it represents strings.
 #'
-#' Basically, \R has a very simple encoding marking mechanism,
+#' Overall, \R has a very simple encoding marking mechanism,
 #' see \code{\link{stri_enc_mark}}. There is an implicit assumption
 #' that your platform's default (native) encoding always extends
 #' ASCII -- \pkg{stringi} checks that whenever your native encoding